+++ provider.tf
@@ -1,9 +1,9 @@
 terraform {
-  required_version = "~> 1.0"
+  required_version = ">= 1.6"
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 4.78.0"
+      version = "~> 4.0"
     }
     google-beta = {
       source  = "hashicorp/google-beta"