+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.14.2] - ????-??-??
+## [0.15.0] - 2023-10-16
 ### Added
 - [Custom input validation rules](https://developer.hashicorp.com/terraform/language/values/variables#custom-validation-rules)
+- [Terrafom Tests](./tests)
+- `hashicorp/random` provider `v3.5.1`
 ### Changed
+- Updated Terraform `required_version` from `~> 1.0` to `>= 1.6`
+- Updated `google` provider from `>= 4.78.0` to `~> 4.0`
 
 ## [0.14.1] - 2023-09-18
 ### Added